spacetime version working too
