+++ plugins/dst/src/patterns.js
